feat: small change of structure, move out pages from /components to /pages
