+++ fe/src/pages/logs/logs.scss
